+++ gui/PotionsListWidget.h
@@ -37,8 +37,8 @@ public slots:
 	void refreshList();
 
 protected:
-	void nbCraftable(int id, QString& text, QString& tooltip);
-	void nbDiscoveredEffects(int id, QString& text, QString& tooltip);
+	void nbCraftable(size_t id, QString& text, QString& tooltip);
+	void nbDiscoveredEffects(size_t id, QString& text, QString& tooltip);
 
 	QColor m_positiveColor, m_negativeColor;
 };
